Add additional agent details with Box components
